Moved signUpForm into Navigation
