+++ frontend/src/App.tsx
@@ -1,94 +1,21 @@
-import { useContext, useEffect, useState } from "react";
+import { useContext } from "react";
 import { Navigate, Route, Routes } from "react-router-dom";
 import Navigation from "./components/Navigation";
-import { NavigationItem } from "./enums/navigation";
-import SignUpForm from "./components/SignUpForm";
 import DebugTimeline from "./routes/DebugTimeline";
 import ProjectCreation from "./routes/ProjectCreation";
 import ProjectCards from "./components/ProjectCards";
-import { UserContext, UserUpdateContext } from "./UserContext";
+import { UserContext } from "./UserContext";
 import Welcome from "./components/Welcome";
 import ManagerFooter from "./components/ManagerFooter";
 import ManageUser from "./routes/ManageUser";
 import { UserRole } from "./interfaces/UserObj";
 
 function App() {
-    const [username, setUsername] = useState<string>("");
-    const [password, setPassword] = useState<string>("");
-    const [isLoginInvalid, setIsLoginInvalid] = useState<boolean>(false);
-
     const userInfo = useContext(UserContext);
-    const updateUserInfo = useContext(UserUpdateContext);
-
-    useEffect(() => {
-        if (document.cookie.startsWith("token=") && userInfo === null) {
-            updateUserInfo();
-        }
-    }, []);
-
-    const login = async (username: string, password: string) => {
-        const url = (import.meta.env.VITE_API_URL as string) + "api/auth/signin";
-
-        const response = await fetch(url, {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json",
-            },
-            body: JSON.stringify({
-                username: username,
-                password: password,
-            }),
-        });
-
-        if (response.ok) {
-            const json = await response.json();
-
-            const date = new Date();
-            date.setTime(date.getTime() + 15 * 60 * 1000);
-            document.cookie = `token=${json.tokenType} ${json.accessToken};expires="${date.toUTCString()};SameSite=Strict;path=/`;
-
-            updateUserInfo();
-        } else {
-            setIsLoginInvalid(true);
-            setPassword("");
-        }
-
-        console.log(url, response.ok, response.status);
-    };
-
-    const singUp = async (username: string, password: string) => {
-        console.log("username", username, "password", password);
-
-        setUsername(username);
-        setPassword(password);
-    };
-
-    const signOut = async () => {
-        console.log("signOut");
-        document.cookie = "token=;expires=Thu, 01 Jan 1970 00:00:00 GMT";
-        updateUserInfo();
-    };
 
     return (
         <div>
-            <Navigation
-                activeNavigationItem={NavigationItem.HOME}
-                onLogin={login}
-                onSignOut={signOut}
-                onSignUp={singUp}
-                username={username}
-                password={password}
-                setPassword={setPassword}
-                setUsername={setUsername}
-                isLoginInvalid={isLoginInvalid}
-            />
-            <SignUpForm
-                signin={login}
-                username={username}
-                password={password}
-                setPassword={setPassword}
-                setUsername={setUsername}
-            ></SignUpForm>
+            <Navigation />
             <Routes>
                 <Route
                     path="/"
